requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=0.1.0

# FastAPI and web dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Database dependencies
psycopg2-binary>=2.9.0
asyncpg>=0.29.0


#DPSN client sdk
dpsn-client

# Environment and configuration
python-dotenv>=1.0.0

# Token counting and utilities
tiktoken>=0.5.0

# AWS S3 integration
boto3>=1.34.0

# HTTP requests
requests>=2.31.0

# JSON and data processing
jsonschema>=4.20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
python-multipart>=0.0.6
